refactor plugin store helpers
