+++ src/ui_logic/utils/api.py
@@ -498,28 +498,6 @@ def fetch_user_workflows(
     except Exception:
         return []
 
-def fetch_store_plugins(
-    limit: int = 50, token: Optional[str] = None, org: Optional[str] = None
-) -> List[Dict[str, Any]]:
-    """Return plugin metadata from the store API or an empty list on error."""
-    try:
-        params = {"limit": limit}
-        return api_get("plugins/store", params=params, token=token, org=org)
-    except Exception:
-        return []
-
-def search_plugins(
-    query: str,
-    limit: int = 50,
-    token: Optional[str] = None,
-    org: Optional[str] = None,
-) -> List[Dict[str, Any]]:
-    """Search public plugin marketplace."""
-    try:
-        params = {"q": query, "limit": limit}
-        return api_get("plugins/search", params=params, token=token, org=org)
-    except Exception:
-        return []
 
 def create_workflow(user_id: str, workflow: Dict[str, Any], token: Optional[str] = None, org: Optional[str] = None) -> bool:
     """Create a new workflow for ``user_id``."""
@@ -716,8 +694,6 @@ __all__ = [
     "fetch_knowledge_graph",
     # Plugins
     "fetch_user_workflows",
-    "fetch_store_plugins",
-    "search_plugins",
     "create_workflow",
     "delete_workflow",
     "update_workflow",
